+++ packages/components/src/tools-panel/stories/utils/tools-panel-with-item-group-slot.tsx
